Move verifiers -> jwt
