+++ src/fastmcp/server/auth/providers/bearer.py
@@ -1,14 +1,14 @@
 """Backwards compatibility shim for BearerAuthProvider.
 
-The BearerAuthProvider class has been moved to fastmcp.server.auth.verifiers.JWTVerifier
+The BearerAuthProvider class has been moved to fastmcp.server.auth.providers.jwt.JWTVerifier
 for better organization. This module provides a backwards-compatible import.
 """
 
 import warnings
 
 import fastmcp
-from fastmcp.server.auth.verifiers import JWKData, JWKSData, RSAKeyPair
-from fastmcp.server.auth.verifiers import JWTVerifier as BearerAuthProvider
+from fastmcp.server.auth.providers.jwt import JWKData, JWKSData, RSAKeyPair
+from fastmcp.server.auth.providers.jwt import JWTVerifier as BearerAuthProvider
 
 # Re-export for backwards compatibility
 __all__ = ["BearerAuthProvider", "RSAKeyPair", "JWKData", "JWKSData"]
@@ -18,7 +18,7 @@ if fastmcp.settings.deprecation_warnings:
     warnings.warn(
         "The `fastmcp.server.auth.providers.bearer` module is deprecated "
         "and will be removed in a future version. "
-        "Please use `fastmcp.server.auth.verifiers.JWTVerifier` "
+        "Please use `fastmcp.server.auth.providers.jwt.JWTVerifier` "
         "instead of this module's BearerAuthProvider.",
         DeprecationWarning,
         stacklevel=2,
